+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Remove namespace and help message prefix for Prometheus exporter. This could be
   a breaking change if you have Prometheus metrics from OpenCensus Prometheus exporter
   of previous versions, please point to the new metrics with no namespace instead.
+- Add an util artifact `opencensus-contrib-appengine-standard-util` to interact with the AppEngine
+  CloudTraceContext.
 
 ## 0.13.2 - 2018-05-08
 - Map http attributes to Stackdriver format (fix [#1153](https://github.com/census-instrumentation/opencensus-java/issues/1153)).